Remove the CAHNRS School of Design and Construction term name
No need to add this to the `get_university_term_updates()` array,
as the commit that introduced it has not been deployed.

This does, however, reveal a potential inadequacy in the current
term removal processing, in that it does not account for cases
where the same term name might twice in the same taxonomy (under
two different parent terms).